Fix issue preventing ollama service from being exposed on Internet
Move ollama and webui containers to a public bridge network, expose port 7861 to allow Internet connections to the service, and set host name to container name for communication.

Adding the containers to a bridge network is best practice. The port 11434 does not need to be exposed to the Internet. Finally, the port 7861 does need to be exposed.

In addition, the host firewall will need to have IP Masquerading turned on to allow the ollama container to connect to the model registry, and the host's ethernet device needs to be added to the "devices" field for the public firewalld zone. I will propose a separate change in the README for these steps.